mistral AI/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.8
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
MarkupSafe==3.0.2
mistralai==0.4.2
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.63.0
orjson==3.10.15
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
twilio==9.4.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
